Restore shell-invoked / python unit test framework
This adds back the "run_tests python" option, and provides a
simple sample test (invoke_wrapper.sh) that checks that the
solar_capture front-end can start a capture.

Note that this test requires a SolarCapture capable card be present.

The result of "make run" should now be 170 tests, all passing.