SolarCapture warns when running on a different version of Onload
SolarCapture is statically linked to ef_vi.

It emits a non-fatal error if you're running against a verison of
Onload older than the one you built on.

It emits a warning if you are on something with a newer major version
(e.g. build on Onload 8.1.3 run on Onload 9.x)

And it emits an info mesage if the versions differ in any other way.

Also includes the ef_vi version compiled against in the initial
startup banner, since it can be useful for debugging.

The version check does allow (and ignores) suffixes, allowing the
solar_capture version check when running on an Onload built from a git
branch rather than only on a release version.